requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
httpx>=0.27
Pillow>=10.3
apify-client>=1.7
pydantic>=2.7
pydantic-settings>=2.2
numpy
imagehash
scikit-learn
opencv-python
python-dotenv
tenacity
torch>=2.1
open-clip-torch==2.24.0
pytesseract
onnxruntime-gpu ; platform_system == "Windows"
onnxruntime     ; platform_system != "Windows"

# OCR detectors

easyocr==1.7.1       
orjson>=3.10.7